feat!: dropped support for Python 3.7
Python 3.7 reached its end-of-life date in June 2023, which means
it will no longer receive bug fixes or security updates from the
Python development team. As a result, we have removed support for
Python 3.7 in order to ensure the ongoing security and stability
of our package. Users who require Python 3.7 can continue to use
older package versions that support it.